Restructure backend routes
The backend routes got entangled, especially the models.py had too much
content for it's name. Split the simulation endpoints into a separate
one and put function that both simulation and grid-simulation use into a
_helpers file for reusing.
